+++ changelog.md
@@ -1,7 +1,10 @@
 ### 0.9.6
 
-- bump bioimageio.spec library version to 0.5.6.0
+- bump bioimageio.spec library version to 0.5.7.1
 - increase default reprducibility tolerance
+- unify quantile (vs percentile) variable names
+- add quantile computation method parameter
+- accept `SampleQuantile` or `DatasetQuantile` as `min`/`max` arguments to `proc_ops.Clip`
 
 ### 0.9.5